+++ CHANGELOG.md
@@ -1,5 +1,9 @@
  Luminati Proxy manager - Change Log
 
+## 1.55.635
+- :star: added back button in stats
+- :star: windows binary file show dialog on port conflicts
+
 ## 1.55.53
 - :bug: fixed high memory usage on high number of concurrent requests